@@ -1,9 +1,7 @@
 class ApplicationController < ActionController::Base
-  # Prevent CSRF attacks by raising an exception.
-  # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
 
-  helper_method :owner?, :body_as_string, :headers_as_string
+  helper_method :bucket, :owner?, :body_as_string, :headers_as_string
 
   before_action :configure_permitted_parameters, if: :devise_controller?
 
@@ -64,7 +62,7 @@ class ApplicationController < ActionController::Base
   end
 
   def bucket
-    @bucket ||= CreateOrRetrieveBucket.call!(
+    @_bucket ||= CreateOrRetrieveBucket.call!(
       token: params[:token],
       owner_token: owner_token,
       user_id: (current_user.id if user_signed_in?)
@@ -72,10 +70,8 @@ class ApplicationController < ActionController::Base
   end
 
   def configure_permitted_parameters
-    devise_parameter_sanitizer.for(:sign_up) { |u| u.permit(:name, :email, :password, :password_confirmation, :remember_me) }
-    devise_parameter_sanitizer.for(:sign_in) { |u| u.permit(:name, :email, :password, :remember_me) }
-    devise_parameter_sanitizer.for(:account_update) do  |u|
-      u.permit(:name, :email, :password, :password_confirmation, :current_password)
-    end
+    devise_parameter_sanitizer.permit(:sign_up, keys: [:name, :email, :password, :password_confirmation, :remember_me])
+    devise_parameter_sanitizer.permit(:sign_in, keys: [:name, :email, :password, :remember_me])
+    devise_parameter_sanitizer.permit(:account_update, keys: [:name, :email, :password, :password_confirmation, :current_password])
   end
 end
